+++ lib/flop/builder.ex
@@ -33,84 +33,102 @@ defmodule Flop.Builder do
   ]
 
   def filter(query, schema_struct, filters, extra_opts) do
-    {query, conditions} =
-      Enum.reduce(
-        filters,
-        {query, true},
-        &apply_filter(&1, schema_struct, extra_opts, &2)
-      )
-
-    where(query, ^conditions)
+    Enum.reduce(
+      filters,
+      query,
+      &apply_filter(&2, &1, schema_struct, extra_opts)
+    )
   end
 
-  defp apply_filter(%Filter{field: nil}, _, _, result), do: result
-  defp apply_filter(%Filter{value: nil}, _, _, result), do: result
+  defp apply_filter(query, %Filter{field: nil}, _, _), do: query
+  defp apply_filter(query, %Filter{value: nil}, _, _), do: query
 
   defp apply_filter(
+         query,
          %Filter{field: field} = filter,
          schema_struct,
-         extra_opts,
-         {query, conditions}
+         extra_opts
        ) do
     case get_field_type(schema_struct, field) do
       {:custom, %{} = custom_opts} ->
         {mod, fun, opts} = Map.fetch!(custom_opts, :filter)
         opts = Keyword.merge(extra_opts, opts)
-
-        {apply(mod, fun, [query, filter, opts]), conditions}
+        apply(mod, fun, [query, filter, opts])
 
       field_type ->
-        {query, build_op(conditions, schema_struct, field_type, filter)}
+        where(query, ^build_op(schema_struct, field_type, filter))
     end
   end
 
-  defp build_op(c, schema_struct, {:compound, fields}, %Filter{op: op} = filter)
-       when op in [
-              :=~,
-              :like,
-              :not_like,
-              :like_and,
-              :like_or,
-              :ilike,
-              :not_ilike,
-              :ilike_and,
-              :ilike_or,
-              :not_empty
-            ] do
-    compound_dynamic =
-      fields
-      |> Enum.map(&get_field_type(schema_struct, &1))
-      |> Enum.reduce(false, fn field, dynamic ->
-        dynamic_for_field =
-          build_op(true, schema_struct, field, %{filter | field: field})
+  for op <- [:like_and, :like_or, :ilike_and, :ilike_or] do
+    {field_op, combinator} =
+      case op do
+        :ilike_and -> {:ilike, :and}
+        :ilike_or -> {:ilike, :or}
+        :like_and -> {:like, :and}
+        :like_or -> {:like, :or}
+      end
 
-        dynamic([r], ^dynamic or ^dynamic_for_field)
+    defp build_op(
+           schema_struct,
+           {:compound, fields},
+           %Filter{op: unquote(op), value: value}
+         ) do
+      fields = Enum.map(fields, &get_field_type(schema_struct, &1))
+
+      value =
+        case value do
+          v when is_binary(v) -> String.split(v)
+          v when is_list(v) -> v
+        end
+
+      reduce_dynamic(unquote(combinator), value, fn substring ->
+        Enum.reduce(fields, false, fn field, inner_dynamic ->
+          dynamic_for_field =
+            build_op(schema_struct, field, %Filter{
+              field: field,
+              op: unquote(field_op),
+              value: substring
+            })
+
+          dynamic([r], ^inner_dynamic or ^dynamic_for_field)
+        end)
       end)
-
-    dynamic([r], ^c and ^compound_dynamic)
+    end
   end
 
   defp build_op(
-         c,
+         schema_struct,
+         {:compound, fields},
+         %Filter{op: op} = filter
+       )
+       when op in [:=~, :like, :not_like, :ilike, :not_ilike, :not_empty] do
+    fields
+    |> Enum.map(&get_field_type(schema_struct, &1))
+    |> Enum.reduce(false, fn field, dynamic ->
+      dynamic_for_field =
+        build_op(schema_struct, field, %{filter | field: field})
+
+      dynamic([r], ^dynamic or ^dynamic_for_field)
+    end)
+  end
+
+  defp build_op(
          schema_struct,
          {:compound, fields},
          %Filter{op: :empty} = filter
        ) do
-    compound_dynamic =
-      fields
-      |> Enum.map(&get_field_type(schema_struct, &1))
-      |> Enum.reduce(true, fn field, dynamic ->
-        dynamic_for_field =
-          build_op(true, schema_struct, field, %{filter | field: field})
+    fields
+    |> Enum.map(&get_field_type(schema_struct, &1))
+    |> Enum.reduce(true, fn field, dynamic ->
+      dynamic_for_field =
+        build_op(schema_struct, field, %{filter | field: field})
 
-        dynamic([r], ^dynamic and ^dynamic_for_field)
-      end)
-
-    dynamic([r], ^c and ^compound_dynamic)
+      dynamic([r], ^dynamic and ^dynamic_for_field)
+    end)
   end
 
   defp build_op(
-         c,
          _schema_struct,
          {:compound, _fields},
          %Filter{op: op, value: _value} = _filter
@@ -134,24 +152,23 @@ defmodule Flop.Builder do
       "Flop: Operator '#{op}' not supported for compound fields. Ignored."
     )
 
-    c
+    true
   end
 
-  defp build_op(c, %module{}, {:normal, field}, %Filter{op: op, value: value})
+  defp build_op(%module{}, {:normal, field}, %Filter{op: op, value: value})
        when op in [:empty, :not_empty] do
     ecto_type = module.__schema__(:type, field)
     value = value in [true, "true"]
     value = if op == :not_empty, do: !value, else: value
 
     case array_or_map(ecto_type) do
-      :array -> dynamic([r], ^c and empty(:array) == ^value)
-      :map -> dynamic([r], ^c and empty(:map) == ^value)
-      :other -> dynamic([r], ^c and empty(:other) == ^value)
+      :array -> dynamic([r], empty(:array) == ^value)
+      :map -> dynamic([r], empty(:map) == ^value)
+      :other -> dynamic([r], empty(:other) == ^value)
     end
   end
 
   defp build_op(
-         c,
          _schema_struct,
          {:join, %{binding: binding, ecto_type: ecto_type, field: field}},
          %Filter{op: op, value: value}
@@ -161,9 +178,9 @@ defmodule Flop.Builder do
     value = if op == :not_empty, do: !value, else: value
 
     case array_or_map(ecto_type) do
-      :array -> dynamic([{^binding, r}], ^c and empty(:array) == ^value)
-      :map -> dynamic([{^binding, r}], ^c and empty(:map) == ^value)
-      :other -> dynamic([{^binding, r}], ^c and empty(:other) == ^value)
+      :array -> dynamic([{^binding, r}], empty(:array) == ^value)
+      :map -> dynamic([{^binding, r}], empty(:map) == ^value)
+      :other -> dynamic([{^binding, r}], empty(:other) == ^value)
     end
   end
 
@@ -171,7 +188,6 @@ defmodule Flop.Builder do
     {fragment, prelude, combinator} = op_config(op)
 
     defp build_op(
-           c,
            _schema_struct,
            {:normal, field},
            %Filter{op: unquote(op), value: value}
@@ -181,7 +197,6 @@ defmodule Flop.Builder do
     end
 
     defp build_op(
-           c,
            _schema_struct,
            {:join, %{binding: binding, field: field}},
            %Filter{op: unquote(op), value: value}