+++ lib/flop/misc.ex
@@ -26,39 +26,4 @@ defmodule Flop.Misc do
       ["%borscht%", "%batchoy%", "%gumbo%"]
   """
   def split_search_text(s), do: s |> String.split() |> Enum.map(&add_wildcard/1)
-
-  @doc """
-  Takes a string representation of a Ecto dynamic query and a string
-  representation of a binding list and returns a quoted dynamic.
-
-      iex> quote_dynamic(nil, "[borscht: b]")
-      nil
-
-      iex> d = "dynamic(<<<binding>>>, field(b, ^field) == ^value)"
-      iex> quote_dynamic(d, "[borscht: b]")
-      {
-        :dynamic,
-        [line: 1],
-        [
-          [borscht: {:b, [line: 1], nil}],
-          {
-            :==,
-            [line: 1],
-            [
-              {:field, [line: 1],
-               [{:b, [line: 1], nil}, {:^, [line: 1], [{:field, [line: 1], nil}]}]},
-              {:^, [line: 1], [{:value, [line: 1], nil}]}
-            ]
-          }
-        ]
-      }
-  """
-  def quote_dynamic(nil, _), do: nil
-
-  def quote_dynamic(dynamic_builder, binding)
-      when is_binary(dynamic_builder) do
-    dynamic_builder
-    |> String.replace("<<<binding>>>", binding)
-    |> Code.string_to_quoted!()
-  end
 end